Style clean up and slightly modify config loading.
Why? We want nodule config to take over more shared config loading logic between
our gateways. I think the pathforward is to adapt config loading more
similar to how microcosm handles its object graph, ultimately relying on the
injector di feature added in styx more heavily than we do today.

This pr: move some config code from our private code to nodule-config and clean
things up a bit.
