+++ circle.yml
@@ -5,7 +5,7 @@ machine:
   pre:
     - mkdir ~/.yarn-cache
   node:
-    version: 6.9.5
+    version: 8.9.1
 
 dependencies:
   cache_directories:
