+++ src/configMaker.js
@@ -1,5 +1,8 @@
-import { get, merge, set } from "lodash";
+import { merge, set } from "lodash";
 import { camelCase } from "./helpers";
+
+import { getInjector } from './graph';
+import { Metadata } from './metadata';
 
 const SEPARATOR = "__";
 
@@ -21,48 +24,30 @@ const makeConfig = (vars) => {
   return merge({}, ...configObjectsArray);
 };
 
-/* Build configuration by merging defaults sections.
- */
-const mergeConfigSections = (metadata, sections) => {
-    return Object.keys(sections).reduce(
-        (acc, key) => {
-            const section = {};
-            section[key] = sections[key](metadata);
-            return merge(acc, section);
-        },
-        metadata,
-    );
-};
-
-
 /* Build the full application configuration.
  */
-const buildConfig = (defaults, vars) => {
-    const injector = getInjector();
+const buildConfig = (name, defaults, vars, debug = false, testing = false) => {
+  const injector = getInjector();
 
-    // load environment variables
-    const environ = makeConfig(vars);
+  // load environment variables
+  const environ = makeConfig(vars);
 
-    // generate metadata
-    const metadata = {
-        environment: get(environ, 'environment', 'dev'),
-        ip: get(environ, 'ip', '0.0.0.0'),
-        port: Number(get(environ, 'port', 3006)),
-    };
+  // generate metadata
+  const metadata = new Metadata(
+    name,
+    debug,
+    testing,
+  );
 
-    const config = merge(
-        // generate configuration defaults (using metadata)
-        mergeConfigSections(metadata, defaults),
-        // override with environment settings
-        environ,
-    );
+  const config = merge(defaults, environ);
+  config.metadata = metadata;
 
-    // save config and return
-    injector.factory('config', () => config);
-    return config;
+  // save config and return
+  injector.factory('config', () => config);
+  return config;
 };
 
 module.exports = {
-    buildConfig,
-    makeConfig,
+  buildConfig,
+  makeConfig,
 };
